Escape multiline comments in the properties with a hash-sign
Some descriptions of properties contain html code, that gets translated
into multiline strings. Those strings are not starting with a hash-sign
and will generate invalid properties files.

Introduce a `comment` function, that can be used by the template engine
to escape these multiline strings into valid multiline strings, where
each line starts with a hash-sign. The comment function strips carriage
returns chars, as in my experiments the resulting property files had no
such chars at other line endings.

As the `comment` function has to be placed in the model of the template
execution, the properties will be placed inside of the `templateVariables`
instead of replacing them completely. That leads to changes in the templates,
where references to `properties` have to be changed to `model.properties`.